pyproject.toml
[project]
name = "wide-language-index"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiohttp",
  "aiofiles",
  "pyquery",
  "sh",
  "pydub",
  "numpy",
  "requests",
  "jsonschema",
  "humanize",
  "feedparser",
  "click",
  "clint",
  "google-api-python-client",
  "characteristic",
  "python-dateutil",
  "boto3>=1.36.16",
  "python-dotenv>=1.0.1",
  "yt-dlp>=2025.1.26",
]

[dependency-groups]
dev = ["pyright>=1.1.393", "ruff>=0.9.5"]

[tool.ruff.lint]
ignore = ["E741"]

[project.scripts]
add-sample = "wide_language_index.add_sample:main"
annotate = "wide_language_index.annotate:main"
annotation-stats = "wide_language_index.annotation_stats:annotation_stats"
audit = "wide_language_index.audit:main"
fetch-index = "wide_language_index.fetch_index:fetch_index"
fetch-rss = "wide_language_index.fetch_rss_feed:main"
generate-clips = "wide_language_index.generate_clips:make_clips"
mirror = "wide_language_index.mirror:main"
normalize = "wide_language_index.normalize:normalize_json_files"
play-offset = "wide_language_index.play_offset:play_offset_cmd"
recode-language = "wide_language_index.recode_language:main"
recode-sample = "wide_language_index.recode_sample:main"

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["wide_language_index*"]
namespaces = false
